requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-docx
python-dotenv
pydantic-settings
librosa
numpy
scipy
soundfile
python-multipart
requests
